continued working with the recipe component. Currently need to debug since the content is no longer rendering. This happened after adding bootstrap for the first time
