+++ src/components/recipe/import.js
@@ -1,0 +1,8 @@
+import foodImage1 from "../../../../assets/foodImage1";
+import foodImage2 from "../../../../assets/foodImage2";
+
+
+export {
+    foodImage1,
+    foodImage2,
+}
